backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
python-dotenv

passlib[bcrypt]
python-multipart
